MSTestをNUnitに置き換えた
