+++ NUnitTest/HatsukoiEye4Test.cs
@@ -1,10 +1,12 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿using NUnit.Framework;
 using PaizaOnlineHackathon8;
 
 namespace UnitTestProject {
-  [TestClass]
+
+  [TestFixture]
   public class HatsukoiEye4Test : POHTest<HatsukoiEye4> {
-    [TestMethod]
+
+    [Test]
     public void TestMethod1() {
       Input.AppendLine("5");
       Input.AppendLine("2 3 5 4 1");
@@ -12,7 +14,7 @@ namespace UnitTestProject {
       Test();
     }
 
-    [TestMethod]
+    [Test]
     public void TestMethod2() {
       Input.AppendLine("7");
       Input.AppendLine("27 100 83 2 57 71 40");
