+++ NUnitTest/HatsukoiSpecial6Test.cs
@@ -1,10 +1,12 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿using NUnit.Framework;
 using PaizaOnlineHackathon8;
 
 namespace UnitTestProject {
-  [TestClass]
+
+  [TestFixture]
   public class HatsukoiSpecial6Test : POHTest<HatsukoiSpecial6> {
-    [TestMethod]
+
+    [Test]
     public void TestMethod1() {
       Input.AppendLine("100 110");
       Input.AppendLine("20 200");
@@ -13,7 +15,7 @@ namespace UnitTestProject {
       Test();
     }
 
-    [TestMethod]
+    [Test]
     public void TestMethod2() {
       Input.AppendLine("980 200");
       Input.AppendLine("200 1000");
