+++ NUnitTest/NUnitTest.csproj
@@ -1,21 +1,16 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{3D4A8698-0259-40F5-9CEF-9FAA6EEAB204}</ProjectGuid>
+    <ProjectGuid>{193B5D69-D247-44FC-B2C9-F54F6487F9A5}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>UnitTestProject</RootNamespace>
-    <AssemblyName>UnitTestProject</AssemblyName>
+    <RootNamespace>NUnitTest</RootNamespace>
+    <AssemblyName>NUnitTest</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
-    <ProjectTypeGuids>{3AC096D0-A1C2-E12C-1390-A8335801FDAB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
-    <VisualStudioVersion Condition="'$(VisualStudioVersion)' == ''">10.0</VisualStudioVersion>
-    <VSToolsPath Condition="'$(VSToolsPath)' == ''">$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)</VSToolsPath>
-    <ReferencePath>$(ProgramFiles)\Common Files\microsoft shared\VSTT\$(VisualStudioVersion)\UITestExtensionPackages</ReferencePath>
-    <IsCodedUITest>False</IsCodedUITest>
-    <TestProjectType>UnitTest</TestProjectType>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <DebugSymbols>true</DebugSymbols>
@@ -56,21 +51,14 @@
       <Private>False</Private>
     </Reference>
     <Reference Include="System" />
+    <Reference Include="System.Core" />
+    <Reference Include="System.Xml.Linq" />
+    <Reference Include="System.Data.DataSetExtensions" />
+    <Reference Include="Microsoft.CSharp" />
+    <Reference Include="System.Data" />
+    <Reference Include="System.Net.Http" />
+    <Reference Include="System.Xml" />
   </ItemGroup>
-  <Choose>
-    <When Condition="('$(VisualStudioVersion)' == '10.0' or '$(VisualStudioVersion)' == '') and '$(TargetFrameworkVersion)' == 'v3.5'">
-      <ItemGroup>
-        <Reference Include="Microsoft.VisualStudio.QualityTools.UnitTestFramework, Version=10.1.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a, processorArchitecture=MSIL" />
-      </ItemGroup>
-    </When>
-    <Otherwise>
-      <ItemGroup>
-        <Reference Include="Microsoft.VisualStudio.QualityTools.UnitTestFramework">
-          <Private>False</Private>
-        </Reference>
-      </ItemGroup>
-    </Otherwise>
-  </Choose>
   <ItemGroup>
     <Compile Include="HatsukoiClothes2Test.cs" />
     <Compile Include="HatsukoiClothes3Test.cs" />
@@ -84,19 +72,19 @@
     <Compile Include="HatsukoiSpecial2Test.cs" />
     <Compile Include="HatsukoiSpecial5Test.cs" />
     <Compile Include="HatsukoiSpecial6Test.cs" />
-    <Compile Include="POHOne\OneHatsukoiSpecial6Test.cs" />
-    <Compile Include="POHOne\OneHatsukoiSpecial5Test.cs" />
-    <Compile Include="POHOne\OneHatsukoiSpecial2Test.cs" />
-    <Compile Include="POHOne\OneHatsukoiClothes3Test.cs" />
-    <Compile Include="POHOne\OneHatsukoiClothes2Test.cs" />
-    <Compile Include="POHOne\OneHatsukoiEye4Test.cs" />
-    <Compile Include="POHOne\OneHatsukoiEye3Test.cs" />
-    <Compile Include="POHOne\OneHatsukoiEye2Test.cs" />
-    <Compile Include="POHOne\OneHatsukoiHair5Test.cs" />
-    <Compile Include="POHOne\OneHatsukoiHair4Test.cs" />
-    <Compile Include="POHOne\OneHatsukoiHair3Test.cs" />
-    <Compile Include="POHOne\OneHatsukoiHair2Test.cs" />
-    <Compile Include="POHOne\OnePOHTest.cs" />
+    <Compile Include="OneLiners\OneHatsukoiClothes2Test.cs" />
+    <Compile Include="OneLiners\OneHatsukoiClothes3Test.cs" />
+    <Compile Include="OneLiners\OneHatsukoiEye2Test.cs" />
+    <Compile Include="OneLiners\OneHatsukoiEye3Test.cs" />
+    <Compile Include="OneLiners\OneHatsukoiEye4Test.cs" />
+    <Compile Include="OneLiners\OneHatsukoiHair2Test.cs" />
+    <Compile Include="OneLiners\OneHatsukoiHair3Test.cs" />
+    <Compile Include="OneLiners\OneHatsukoiHair4Test.cs" />
+    <Compile Include="OneLiners\OneHatsukoiHair5Test.cs" />
+    <Compile Include="OneLiners\OneHatsukoiSpecial2Test.cs" />
+    <Compile Include="OneLiners\OneHatsukoiSpecial5Test.cs" />
+    <Compile Include="OneLiners\OneHatsukoiSpecial6Test.cs" />
+    <Compile Include="OneLiners\OnePOHTest.cs" />
     <Compile Include="POHTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
@@ -109,25 +97,9 @@
   <ItemGroup>
     <None Include="packages.config" />
   </ItemGroup>
-  <Choose>
-    <When Condition="'$(VisualStudioVersion)' == '10.0' And '$(IsCodedUITest)' == 'True'">
-      <ItemGroup>
-        <Reference Include="Microsoft.VisualStudio.QualityTools.CodedUITestFramework, Version=10.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a, processorArchitecture=MSIL">
-          <Private>False</Private>
-        </Reference>
-        <Reference Include="Microsoft.VisualStudio.TestTools.UITest.Common, Version=10.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a, processorArchitecture=MSIL">
-          <Private>False</Private>
-        </Reference>
-        <Reference Include="Microsoft.VisualStudio.TestTools.UITest.Extension, Version=10.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a, processorArchitecture=MSIL">
-          <Private>False</Private>
-        </Reference>
-        <Reference Include="Microsoft.VisualStudio.TestTools.UITesting, Version=10.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a, processorArchitecture=MSIL">
-          <Private>False</Private>
-        </Reference>
-      </ItemGroup>
-    </When>
-  </Choose>
-  <Import Project="$(VSToolsPath)\TeamTest\Microsoft.TestTools.targets" Condition="Exists('$(VSToolsPath)\TeamTest\Microsoft.TestTools.targets')" />
+  <ItemGroup>
+    <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
