+++ NUnitTest/OneLiners/OneHatsukoiClothes3Test.cs
@@ -1,24 +1,26 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿using NUnit.Framework;
 using PaizaOnlineHackathon8;
 
 namespace UnitTestProject {
-  [TestClass]
+
+  [TestFixture]
   public class OneHatsukoiClothes3Test : OnePOHTest<HatsukoiClothes3_One> {
-    [TestMethod]
+
+    [Test]
     public void TestMethod1() {
       Input.AppendLine("5 2");
       Expected.AppendLine("3");
       Test();
     }
 
-    [TestMethod]
+    [Test]
     public void TestMethod2() {
       Input.AppendLine("10 10");
       Expected.AppendLine("0");
       Test();
     }
 
-    [TestMethod]
+    [Test]
     public void TestMethod3() {
       Input.AppendLine("2 4");
       Expected.AppendLine("0");
