+++ PaizaOnlineHackathon8.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 14.0.24720.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PaizaOnlineHackathon8", "PaizaOnlineHackathon8\PaizaOnlineHackathon8.csproj", "{7B671AA9-CB52-4CB1-8F5A-632BE43173F1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnitTestProject", "UnitTestProject\UnitTestProject.csproj", "{3D4A8698-0259-40F5-9CEF-9FAA6EEAB204}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NUnitTest", "NUnitTest\NUnitTest.csproj", "{193B5D69-D247-44FC-B2C9-F54F6487F9A5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,10 +17,10 @@ Global
 		{7B671AA9-CB52-4CB1-8F5A-632BE43173F1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7B671AA9-CB52-4CB1-8F5A-632BE43173F1}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7B671AA9-CB52-4CB1-8F5A-632BE43173F1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3D4A8698-0259-40F5-9CEF-9FAA6EEAB204}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3D4A8698-0259-40F5-9CEF-9FAA6EEAB204}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3D4A8698-0259-40F5-9CEF-9FAA6EEAB204}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3D4A8698-0259-40F5-9CEF-9FAA6EEAB204}.Release|Any CPU.Build.0 = Release|Any CPU
+		{193B5D69-D247-44FC-B2C9-F54F6487F9A5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{193B5D69-D247-44FC-B2C9-F54F6487F9A5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{193B5D69-D247-44FC-B2C9-F54F6487F9A5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{193B5D69-D247-44FC-B2C9-F54F6487F9A5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
